+++ src/scicloj/metamorph/ml/design_matrix.clj
@@ -168,4 +168,3 @@
         (ds/categorical->number cf/categorical))))
 
 
-